ayomi/requirements.txt
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.1
